+++ Tarea2/Models/Cabecera.cs
@@ -4,8 +4,8 @@
 	{
 		public int ID { get; set; }
 		public string? Nombre { get; set; }
-		public double Longitud { get; set; }
-		public double Latitud { get; set; }
+		public decimal Longitud { get; set; }
+		public decimal Latitud { get; set; }
 		public string? Imagen { get; set; }
 	}
 }
